added multi-server management, renamed some functions
This was a long one in the making, but it's finally here! This bot is now capable of handling multiple Minecraft servers at once! To do this, a lot of other changes needed to be made, such as moving to YAMLs to store data in a more orderly fashion. The chat integration has also been redone to use threads instead. Other than that, the commands relating to the server address have also been renamed to be more intuitive. The bot now saves address changes automatically so you don't need to worry about having to edit the config when restarting.